% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_trapz.R
\name{cdf_trapz}
\alias{cdf_trapz}
\title{Computing the area under curve using trapezoidal method}
\usage{
cdf_trapz(f, a = 0, b, n.pts = 1001)
}
\arguments{
\item{f}{A function of curve.}

\item{a}{A starting point value.}

\item{b}{An ending point value.}

\item{n.pts}{A number of subdivisions of trapezoid shape under the curve.}
}
\value{
A value.
}
\description{
Computing the area under curve using trapezoidal method
}
\examples{
f <- exp(-x^2)
cdf_trapz(f=f, a=0, b=1, n.pts=15)
}
