% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dataPH.R
\name{generate_dataPH}
\alias{generate_dataPH}
\title{Function to generate bivariate data using Proportional Hazard Model using inverse sampling technique}
\usage{
generate_dataPH(mod, params, n.sample)
}
\arguments{
\item{mod}{A string that specify which model to be implemented to join the bivariate data (join biomarker X and time-to-event T).
mod = c('exp_exp', 'norm_exp', 'norm_weib')}

\item{params}{A list of parameters for selected model.
'exp_exp': params = list(c.rate, beta, lambda.t, mu.x).
'norm_exp': params = list(c.rate, beta, lambda.t, mu.x, sigma.x).
'norm_weib': params = list(c.rate, beta, scale.t, shape.t, mu.x, sigma.x).
c.rate: censoring rate (0 to 1)
beta: association measurement
lambda.t: scale parameter of T
scale.t: scale parameter of T (for weibull)
shape.t: shape parameter of T (for weibull)
mu.x: scale parameter of X
sigma.x: standard deviation of X}

\item{n.sample}{A number of sample size to be generates.}
}
\value{
A dataframe with 3 columns (x = biomarker value, t = observable time-to-event, status = censored/not censor (0 or 1))
}
\description{
Function to generate bivariate data using Proportional Hazard Model using inverse sampling technique
}
\examples{
data <- generate_dataPH('exp_exp', params=list(c.rate=0.9, beta=1.2, lambda.t=0.5, mu.x=1), n.sample=200)
data <- generate_dataPH('norm_exp', params=list(c.rate=0.9, beta=0.8, lambda.t=0.5, mu.x=5, sigma.x=0.8), n.sample=200)
data <- generate_dataPH('norm_weib', params=list(c.rate=0.9, beta=0.2, scale.t=1/1.5, shape.t=1.2, mu.x=5, sigma.x=0.8), n.sample=200)
}
