% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCestimate_COP.R
\name{MCMCestimate_COP}
\alias{MCMCestimate_COP}
\title{MCMCestimate_COP}
\usage{
MCMCestimate_COP(
  model,
  df,
  f.path,
  p.jags,
  i.jags,
  n.chains = 3,
  n.iter = 50000,
  n.thin = 1,
  n.burnin = 2000
)
}
\arguments{
\item{model}{A string which specify the model of interest.\cr
model = c('exp_exp','norm_exp','norm_weib').}

\item{df}{A dataframe with 3 columns ('x' = biomarker, 't' = time-to-event, 'status' = censoring status (0 or 1))}

\item{f.path}{A string that specify file path to store value of estimated parameters.}

\item{p.jags}{A vector of parameters that need to be monitored.\cr
'exp_exp':  c('lambda.T','theta','mu.X','Deviance')\cr
'norm_exp': c('lambda.T','theta','mu.X','sigma.X','Deviance')\cr
'norm_weib': c('scale.T','shape.T','theta','mu.X','sigma.X','Deviance')\cr
Deviance: Average of -2*loglikelihood (Dbar) - Will be used for manual DIC calculation.}

\item{i.jags}{A function that generate the initial value of parameter used in JAGS.\cr
'exp_exp': function(){list(lambda.T=lambda.t,theta=0.4, mu.X=mu.x)}\cr
'norm_exp': function(){list(lambda.T=lambda.t,theta=theta, mu.X=mu.x, tau=runif(1))}\cr
'norm_weib': function(){list(k=scale.t,shape.T=shape.t,theta=theta, mu.X=mu.x, tau=runif(1))}\cr
theta: A value of dependence measurement\cr
lambda.T: A scale parameter of T\cr
k: A scale parameter of T (for weibull; JAGS use a different version of scale)\cr
shape.T: A shape parameter of T (for weibull)\cr
mu.x: A scale parameter for X\cr
tau: A standard deviation for X (JAGS suggest using tau)}

\item{n.chains}{A number of chains to run MCMC process.}

\item{n.iter}{A number of iteration run for each chain.}

\item{n.thin}{A number of iteration thinning.}

\item{n.burnin}{A number that specify how many iteration will be discarded.}
}
\value{
Print MCMC summary & automatically save the summary to the specified file path.
}
\description{
Function to run estimation process of joint model discussed in this package using MCMC Gibbs Sampling.
}
\examples{
mu.x=5; sigma.x=0.8; shape.t=1.2; scale.t=1/1.4; tau=0.7; theta=2*tau/(1-tau)
f.path='C:/Users/farea/Documents/RWorkspace'
p.jags <- c('scale.T','shape.T','theta','mu.X','sigma.X','Deviance')
i.jags <- function(){list(k=scale.t,shape.T=shape.t,theta=theta, mu.X=mu.x, tau=runif(1))}
MCMCestimate_COP(model='norm_weib', df=df.COP, f.path=f.path, p.jags=p.jags, i.jags=i.jags)
}
