% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_COP.R
\name{roc_COP}
\alias{roc_COP}
\title{Function to produce time-dependent ROC using Rotated 90 Clayton copula.}
\usage{
roc_COP(max.x, time.t, mod, params, n.quant = 10)
}
\arguments{
\item{max.x}{A value that specify what is the maximum X of our data.}

\item{time.t}{A vector of T that we are interested to produce its ROC.}

\item{mod}{A string that specify our joint model of biomarker X and time-to-event T.
mod = c('exp_exp','norm_exp','norm_weib')}

\item{params}{A list of estimated parameters.
'exp_exp': params = list( theta, lambda.t, mu.x).
'norm_exp': params = list(theta, lambda.t, mu.x, sigma.x).
'norm_weib': params = list(theta, scale.t, shape.t, mu.x, sigma.x).
theta: dependence measurement (tau (0 to 1);theta=2*tau/(1-tau))
lambda.t: scale parameter of T
scale.t: scale parameter of T (for weibull)
shape.t: shape parameter of T (for weibull)
mu.x: scale parameter of X
sigma.x: standard deviation of X}

\item{n.quant}{A number of how many quantile of Specificity that will be used in the ROC curve.}
}
\value{
A dataframe of 4 columns (sensitivity, specificity, time, cutoff.x)
}
\description{
Function to produce time-dependent ROC using Rotated 90 Clayton copula.
}
\examples{
tau=0.5; theta=2*tau/(1-tau); n.quant=21; mu.x=1.4; lambda.t=0.2;
params=list(mu.x=mu.x,lambda.t=lambda.t,theta=theta)
roc_COP(max.x=max(df.COP$x), time.t=t, mod='exp_exp', params=params, n.quant=n.quant)
}
