% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_multisn.R
\name{roc_multisn}
\alias{roc_multisn}
\title{roc_multisn}
\usage{
roc_multisn(
  max.x,
  time.t,
  mod,
  params,
  n.sample = 1e+05,
  cutoff.n = 10001,
  n.quant = 10,
  low.lim = NULL,
  upp.lim = NULL
)
}
\arguments{
\item{max.x}{A value that specify what is the maximum X of our data.}

\item{time.t}{A vector of T that we are interested to produce its ROC.}

\item{mod}{A string that specify our joint model of biomarker X and time-to-event T. \cr
mod = c('exp_exp','norm_exp','norm_weib')}

\item{params}{A list of estimated parameters.\cr
'exp_exp': params = list(beta, lambda.t, mu.x).\cr
'norm_exp': params = list(beta, lambda.t, mu.x, sigma.x).\cr
'norm_weib': params = list(beta, scale.t, shape.t, mu.x, sigma.x).\cr
beta: association measurement\cr
lambda.t: scale parameter of T\cr
scale.t: scale parameter of T (for weibull)\cr
shape.t: shape parameter of T (for weibull)\cr
mu.x: scale parameter of X\cr
sigma.x: standard deviation of X}

\item{n.sample}{A value of sample size that will be used in Monte Carlo Integration method (to find inverse of Specificity).}

\item{cutoff.n}{A value of how many biomarker subdivision is used (to find inverse of Specificity).}

\item{n.quant}{A number of how many quantile of Specificity that will be used in the ROC curve.}
}
\value{
A dataframe of 4 columns (sensitivity, specificity, time, cutoff.x)
}
\description{
Function to produce time-dependent ROC using Proportional Hazard Model.
}
\examples{
beta=1.2; n.quant=21; mu.x=1.4; lambda.t=0.2;
params=list(mu.x=mu.x,lambda.t=lambda.t,beta=beta)
roc_multisn(max.x=max(df.COP$x), time.t=t, mod='exp_exp', params=params, n.quant=n.quant)

}
